+++ chapter-62-magicitems/src/movement_system.rs
@@ -27,7 +27,7 @@ impl<'a> System<'a> for MovementSystem {
         // Apply teleports
         for (entity, teleport) in (&entities, &apply_teleport).join() {
             if teleport.dest_depth == map.depth {
-                apply_move.insert(entity, ApplyMove{ dest_idx: map.xy_idx(teleport.dest_x, teleport.dest_y) as i32 })
+                apply_move.insert(entity, ApplyMove{ dest_idx: map.xy_idx(teleport.dest_x, teleport.dest_y) })
                     .expect("Unable to insert");
             } else if entity == *player_entity {
                 *runstate = RunState::TeleportingToOtherLevel{ x: teleport.dest_x, y: teleport.dest_y, depth: teleport.dest_depth };
@@ -55,8 +55,8 @@ impl<'a> System<'a> for MovementSystem {
                 map.blocked[start_idx] = false;
                 map.blocked[dest_idx] = true;
             }
-            pos.x = movement.dest_idx % map.width;
-            pos.y = movement.dest_idx / map.width;
+            pos.x = movement.dest_idx as i32 % map.width;
+            pos.y = movement.dest_idx as i32 / map.width;
             if let Some(vs) = viewsheds.get_mut(entity) {
                 vs.dirty = true;
             }
